+++ NEWS.md
@@ -1,7 +1,9 @@
 # mmrm 0.2.2.9001
 
 ### New Features
+
 - Add Kenward-Roger support for spatial covariance structures.
+- Add optional tidymodels support (#214, @dgkf)
 
 # mmrm 0.2.2